Move streaming to common
