+++ src/common/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,0 +1,1 @@
+osmesa.common.streaming.ChangesetsProvider
